requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Async support
httpx==0.27.0
aiofiles==23.2.1

# Config
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Embeddings
sentence-transformers==2.7.0

# Vector DB - Qdrant Cloud
qdrant-client==1.9.1

# LLM - Mistral AI
mistralai==1.2.5

# PDF processing
pypdf==4.2.0

# Text splitting
langchain-text-splitters==0.2.0

# Data
numpy==1.26.4

# Logging
loguru==0.7.2

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6

streamlit==1.30.0

# Evaluation
rouge-score==0.1.2
bert-score==0.3.13
ragas==0.1.21
datasets==2.19.0


